+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 - Mise en place du menu
 - Mise en place du header
 - Mise en place du footer
+- Mise en place du breadcrumb